Make download CEF automatically an optional feature
I would like to use this crate, but am not fond of the automatic downloading stuff. I saw that I can specify CEF_PATH, but that will still check a custom archive.json file that I don't have.

This adds a default "download" feature that controls all the download logic (and the 100~ build dependency crates that pulls in...). When not active, CEF_PATH will not be checked for the archive file.
